### 2022-07-16 #4570 @kyleniemeyer
@openjournals/dev it looks like this submission has the paper in an orphaned branch, or at least one that doesn't have any of the actual software package. Will that work for doing the repo checks?
