### 2022-07-16 #4570 @kyleniemeyer
@openjournals/dev or, is there some other reason why the repo details and license are not found? The repo link above looks ok, actually, so I'm not sure what is going on.
