### 2022-07-17 #4570 @xuanxu
> @openjournals/dev or, is there some other reason why the repo details and license are not found? The repo link above looks ok, actually, so I'm not sure what is going on.

The repo checks are by default performed on the branch set in the header. But the command should accept other branch, let's try
