### 2022-07-21 #4570 @dunnkers
Hi @diehlpk. Great that you accepted to be the editor for our submission. We suggest the following reviewers:

- Talavera Martinez, Estefania <e.talaveramartinez@utwente.nl>
- Michael Biehl <m.biehl@rug.nl>
- Antonio GRECO <agreco@unisa.it>
- Manuel Castejon Limas <manuel.castejon@unileon.es>
- leotrs <leo@leotrs.com> (From JOSS reviewers list. Python proficiency and dimensionality reduction amongst domains)
- yxoos <ah7557@rit.edu> (From JOSS reviewers list. Python and feature selection amongst domains)

✓
