### 2022-07-28 9b9d096
Use pd.concat instead of df.append()
