+++ tests/unit/callbacks/test_to_wandb.py
@@ -113,6 +113,14 @@ def test_on_summary(wandb_callback: WandbCallback):
     # → assertions in `test_on_end`
 
 
+@pytest.mark.xfail(
+    "The Weights-and-Biases platform has proven to have too large delays to be able to"
+    + " run integration tests on it consistently. Alhough the WandB platform does"
+    + " eventually upload all the data that is being sent to it, all logging is queued,"
+    + " making it hard to rely on it in integration-/ unit test settings."
+    + "Therefore this test is marked as optional to prevent unnecessary build failures."
+    + " The test is kept in the codebase, however, for convenience in local debugging."
+)
 @pytest.mark.dependency(
     depends=[
         "test_on_begin",
@@ -130,7 +138,6 @@ def test_on_end(wandb_callback: WandbCallback, api: Api, request: FixtureRequest
 
     # `on_end`
     wandb_callback.on_end()
-    sleep(5)  # wait to give wandb run time to end
 
     # ensure finished
     api._runs = {}  # reset wandb runs internal cache
