### 2022-07-28 6a2db4f
Upload examples as zips in GH actions
