+++ assets/js/567f2fe4.76998569.js
@@ -1,1 +1,0 @@
-"use strict";(self.webpackChunkwebsite=self.webpackChunkwebsite||[]).push([[543],{8215:function(t,e,a){var n=a(7294);e.Z=function(t){var e=t.children,a=t.hidden,r=t.className;return n.createElement("div",{role:"tabpanel",hidden:a,className:r},e)}},9877:function(t,e,a){a.d(e,{Z:function(){return p}});var n=a(7462),r=a(7294),l=a(2389),i=a(5773),s=a(6010),o="tabItem_LplD";function d(t){var e,a,l,d=t.lazy,p=t.block,u=t.defaultValue,c=t.values,m=t.groupId,f=t.className,k=r.Children.map(t.children,(function(t){if((0,r.isValidElement)(t)&&void 0!==t.props.value)return t;throw new Error("Docusaurus error: Bad <Tabs> child <"+("string"==typeof t.type?t.type:t.type.name)+'>: all children of the <Tabs> component should be <TabItem>, and every <TabItem> should have a unique "value" prop.')})),g=null!=c?c:k.map((function(t){var e=t.props;return{value:e.value,label:e.label,attributes:e.attributes}})),h=(0,i.lx)(g,(function(t,e){return t.value===e.value}));if(h.length>0)throw new Error('Docusaurus error: Duplicate values "'+h.map((function(t){return t.value})).join(", ")+'" found in <Tabs>. Every value needs to be unique.');var b=null===u?u:null!=(e=null!=u?u:null==(a=k.find((function(t){return t.props.default})))?void 0:a.props.value)?e:null==(l=k[0])?void 0:l.props.value;if(null!==b&&!g.some((function(t){return t.value===b})))throw new Error('Docusaurus error: The <Tabs> has a defaultValue "'+b+'" but none of its children has the corresponding value. Available values are: '+g.map((function(t){return t.value})).join(", ")+". If you intend to show no default tab, use defaultValue={null} instead.");var N=(0,i.UB)(),v=N.tabGroupChoices,y=N.setTabGroupChoices,C=(0,r.useState)(b),_=C[0],w=C[1],T=[],D=(0,i.o5)().blockElementScrollPositionUntilNextRender;if(null!=m){var x=v[m];null!=x&&x!==_&&g.some((function(t){return t.value===x}))&&w(x)}var I=function(t){var e=t.currentTarget,a=T.indexOf(e),n=g[a].value;n!==_&&(D(e),w(n),null!=m&&y(m,n))},S=function(t){var e,a=null;switch(t.key){case"ArrowRight":var n=T.indexOf(t.currentTarget)+1;a=T[n]||T[0];break;case"ArrowLeft":var r=T.indexOf(t.currentTarget)-1;a=T[r]||T[T.length-1]}null==(e=a)||e.focus()};return r.createElement("div",{className:"tabs-container"},r.createElement("ul",{role:"tablist","aria-orientation":"horizontal",className:(0,s.Z)("tabs",{"tabs--block":p},f)},g.map((function(t){var e=t.value,a=t.label,l=t.attributes;return r.createElement("li",(0,n.Z)({role:"tab",tabIndex:_===e?0:-1,"aria-selected":_===e,key:e,ref:function(t){return T.push(t)},onKeyDown:S,onFocus:I,onClick:I},l,{className:(0,s.Z)("tabs__item",o,null==l?void 0:l.className,{"tabs__item--active":_===e})}),null!=a?a:e)}))),d?(0,r.cloneElement)(k.filter((function(t){return t.props.value===_}))[0],{className:"margin-vert--md"}):r.createElement("div",{className:"margin-vert--m
[…]

[…]
