+++ website/zip-examples.sh
@@ -1,0 +1,3 @@
+cd ../examples
+zip -r ../website/static/quick-start-structured-configs quick-start-structured-configs
+zip -r ../website/static/quick-start-yaml quick-start-yaml
