+++ .github/workflows/python-app.yml
@@ -48,22 +48,3 @@ jobs:
       - name: Codecov
         if: matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest'
         uses: codecov/codecov-action@v2.1.0
-      - name: Zip examples
-        if: matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest'
-        run: |
-          mkdir -p /tmp/fseval-examples-zips
-          cd examples
-          zip -r /tmp/fseval-examples-zips/quick-start-structured-configs quick-start-structured-configs
-          zip -r /tmp/fseval-examples-zips/quick-start-yaml quick-start-yaml
-      - name: Upload zip for Quick start - structured configs
-        if: matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest'
-        uses: actions/upload-artifact@v3
-        with:
-          name: quick-start-structured-configs
-          path: /tmp/fseval-examples-zips/quick-start-structured-configs.zip
-      - name: Upload zip for Quick start - yaml
-        if: matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest'
-        uses: actions/upload-artifact@v3
-        with:
-          name: quick-start-yaml
-          path: /tmp/fseval-examples-zips/quick-start-yaml.zip
