+++ assets/js/c179f821.4c82eff2.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunkwebsite=self.webpackChunkwebsite||[]).push([[699],{6160:function(e,t,n){n.r(t),n.d(t,{contentTitle:function(){return E},default:function(){return S},frontMatter:function(){return I},metadata:function(){return T},toc:function(){return A}});var a=n(7462),r=n(3366),i=n(7294),s=n(3905),o=n(9877),l=n(8215),c=n(8906),m=n(1832),f=n(6934),u=n(102),p=(n(4251),function e(t,n,a){void 0===n&&(n={items:{}}),void 0===a&&(a="");for(var r=0,i=Object.entries(t);r<i.length;r++){var s=i[r],o=s[0],l=s[1],c="string"!=typeof l,m=(a+"/"+o).replace("/root",""),f=m.split(".").pop();n.items[o]={index:o,hasChildren:c,children:c?Object.keys(l):void 0,canMove:!1,canRename:!1,data:{filePath:m,fileExtension:f,key:o,content:c?null:l}},c&&e(l,n,a+"/"+o)}return n});function d(e){var t=e.template,n=e.treeId,a=e.viewState,r=e.onPrimaryAction,s=e.environment,o=p(t);return i.createElement(m.q,{dataProvider:new f.j(o.items,(function(e,t){return Object.assign({},e,{data:t})})),getItemTitle:function(e){return e.data.key},viewState:a,onPrimaryAction:r},i.createElement(u.m,{treeId:n,rootItem:"root",treeLabel:"Tree Example",ref:s}))}var _=n(1736),k={data:{filePath:"",fileExtension:"",content:""}};function g(e){var t=e.template,n=e.treeId,a=e.viewState,r=e.defaultItem,s=void 0===r?k:r,o=(0,i.useRef)(),l=(0,i.useState)(s),m=l[0],f=l[1],u=(m||k).data;return(0,i.useEffect)((function(){try{for(var e,t=o.current,a=t.treeEnvironmentContext.viewState[n].selectedItems,r=(0,c.Z)(a);!(e=r()).done;){var i=e.value;i&&t.invokePrimaryAction(i,n)}}catch(s){return}})),i.createElement("div",{className:"row"},i.createElement("div",{className:"col col--4"},i.createElement(d,{template:t,treeId:n,viewState:a,onPrimaryAction:f,environment:o})),i.createElement("div",{className:"col col--8"},i.createElement(_.Z,{className:"language-"+u.fileExtension,title:u.filePath},u.content)))}var h=n(7754),v=n(3221),y=n(7557),b=n(6572),w=n(1262),N=["src"],C=function(e){var t=e.src,a=(0,r.Z)(e,N);return i.createElement(w.Z,null,(function(){var e=n(7011),r=(0,i.useRef)(null);return(0,i.useEffect)((function(){var n=r.current;e.create(t,n,a)}),[t]),i.createElement("div",{ref:r})}))},x=["components"],I={sidebar_position:1},E="Getting started",T={unversionedId:"quick-start",id:"quick-start",title:"Getting started",description:"To get started, there's two main resources.",source:"@site/docs/quick-start.mdx",sourceDirName:".",slug:"/quick-start",permalink:"/fseval/docs/quick-start",editUrl:"https://github.com/dunnkers/fseval/tree/website/docs/quick-start.mdx",tags:[],version:"current",sidebarPosition:1,frontMatter:{sidebar_position:1},sidebar:"tutorialSidebar",previous:{title:"Motivation",permalink:"/fseval/docs/motivation"},next:{title:"The pipeline",permalink:"/fseval/docs/the-pipeline"}},A=[{value:"\u26a1\ufe
[…]

[…]
