### 2022-09-07 #4611 @mcasl
The “Documentation” section of my checklist shows that the “Installation instructions” need improvement. Though there isa requeriments.txt file to automatically install the dependencies, the documentation fails to explicitly describe these dependencies.
