### 2022-09-07 #4611 @mcasl
The checklist shows that “Functionality documentation” needs improvement. Though the methods’ documentation reflects the parameters, results and a succinct explanation of what they do, more time has to be devoted to the documentation to clearly explain the new user the elements of the library and how to use them.
