### 2022-09-17 1281d07
Fix console script entry point 🔨
