### 2022-09-17 31388ad
Add alternative installation option -> installing from source
