### 2022-09-17 f2945e0
Add alternative installation method in Documentation
