### 2022-09-21 1367b54
Use UTF-8 when opening README.md
