### 2022-09-21 54c7232
Improve installation instructions (#81)
* Fix console script entry point 🔨

* Add alternative installation option
-> installing from source

* Add alternative installation method in Documentation

* Pin to Python 3.10.6 in CI
Due to python/mypy#13627

* Use UTF-8 when opening README.md
