+++ .github/workflows/python-app.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10"]
+        python-version: ["3.7", "3.8", "3.9", "3.10.6"]
         os: [ubuntu-latest, macOS-latest, windows-latest]
     name: Python ${{ matrix.python-version }} - ${{ matrix.os }}
 
