### 2022-09-21 #4611 @dunnkers
> @dunnkers how are things going?

Hi @diehlpk. I just merged https://github.com/dunnkers/fseval/pull/81 and am working on https://github.com/dunnkers/fseval/pull/83, making improvements where necessary. So it's in-progress.
