### 2022-09-21 3205082
Algorithm stability recipe 📈 First version
