+++ website/docs/_recipes/running-on-slurm.md
@@ -1,1 +1,0 @@
-# Running on a SLURM cluster
