+++ website/docs/_recipes/running-on-aws.md
@@ -1,3 +1,0 @@
-# Running on AWS
-
-evaluate!
