+++ website/docs/_recipes/algorithm-stability.md
@@ -1,9 +1,0 @@
-# Analyze algorithm stability
-
-To analyze the stability of an algorithm, multiple bootstraps can be run using ![`n_bootstraps`](http://localhost:3000/fseval/docs/config/PipelineConfig/) > 1.
-
-For example:
-
-```yaml
-
-```
