+++ website/docs/recipes/algorithm-stability.md
@@ -1,0 +1,284 @@
+# Analyze algorithm stability
+
+
+.
+
+.
+
+.
+
+.
+
+Hi! Let's analyze the results of the experiment you just ran. To recap:
+
+1. You just ran something similar to:
+
+    `python benchmark.py --multirun ranker="glob(*)" +callbacks.to_sql.url="sqlite:///$HOME/results.sqlite"`
+2. There now should exist a `.sqlite` file at this path: `$HOME/results.sqlite`:
+
+    ```
+    $ ls -al $HOME/results.sqlite
+    -rw-r--r-- 1 vscode vscode 20480 Sep 21 08:16 /home/vscode/results.sqlite
+    ```
+
+Let's now analyze the results! 📈
+
+First, we will install `plotly-express`, so we can make nice plots later.
+
+
+```python
+%pip install plotly-express --quiet
+```
+
+    Note: you may need to restart the kernel to use updated packages.
+
+
+Figure out the SQL connection URI.
+
+
+```python
+import os
+
+con: str = "sqlite:///" + os.environ["HOME"] + "/results.sqlite"
+con
+```
+
+
+
+
+    'sqlite:////home/vscode/results.sqlite'
+
+
+
+Read in the `experiments` table. This table contains metadata for all 'experiments' that have been run.
+
+
+```python
+import pandas as pd
+
+experiments: pd.DataFrame = pd.read_sql_table("experiments", con=con, index_col="id")
+experiments
+```
+
+
+
+
+<div>
+<table border="1" class="dataframe">
+  <thead>
+    <tr >
+      <th></th>
+      <th>dataset</th>
+      <th>dataset/n</th>
+      <th>dataset/p</th>
+      <th>dataset/task</th>
+      <th>dataset/group</th>
+      <th>dataset/domain</th>
+      <th>ranker</th>
+      <th>validator</th>
+      <th>local_dir</th>
+      <th>date_created</th>
+    </tr>
+    <tr>
+      <th>id</th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>38vqcwus</th>
+      <td>Synclf hard</td>
+      <td>10000</td>
+      <td>50</td>
+      <td>classification</td>
+      <td>Synclf</td>
+      <td>synthetic</td>
+      <td>Boruta</td>
+      <td>k-NN</td>
+      <td>/workspaces/fseval/examples/algorithm-stabilit...</td>
+      <td>2022-09-21 08:22:28.965510</td>
+    </tr>
+    <tr>
+      <th>y6bb1hcc</th>
+      <td>Synclf hard</td>
+      <td>1000</td>
+      <td>50</td>
+      <td>classification</td>
+      <td>Synclf</td>
+      <td>synthetic</td>
+      <td>Boruta</td>
+      <td>k-NN</td>
+      <td>/workspaces/fseval/examples/algorithm-stabilit...</td>
+      <td>2022-09-21 08:22:53.609396</td>
+    </tr>
+    <tr>
+      <th>3vtr13pg</th>
+      <td>Synclf hard</td>
+      <td>1000</td>
+      <td>50</td>
+      <td>classification</td>
+      <td>Synclf</td>
+      <td>synthetic</td>
+      <td>ReliefF</td>
+      <td>k-NN</td>
+      <td>/workspaces/fseval/examples/algorithm-stabilit...</td>
+      <td>2022-09-21 08:25:09.974370</td>
+    </tr>
+  </tbody>
+</table>
+</div>
+
+
+
+That's looking good 🙌🏻.
+
[…]
