+++ examples/algorithm-stability-yaml/conf/dataset/synclf_hard.yaml
@@ -11,7 +11,7 @@ adapter:
   n_informative: 4
   n_redundant: 0
   n_repeated: 0
-  n_samples: 100
+  n_samples: 1000
   random_state: 0
   shuffle: false
 feature_importances:
