### 2022-09-28 #4611 @estefaniatalavera
Hello,

Can you please change the contact email to the one from my new employer? The old one has been deactivated.

***@***.*** ***@***.***>

Thank you!
Best regards,
Estefania

——————————————————————————
Dr. Estefanía Talavera | Assistant Professor
Data Management and Biometrics | University of Twente
Zilverling room 4098
https://estefaniatalavera.github.io/ <https://estefaniatalavera.github.io/> | https://people.utwente.nl/e.talaveramartinez <https://people.utwente.nl/e.talaveramartinez>

> On 23 Aug 2022, at 17:35, The Open Journals editorial robot ***@***.*** ***@***.***>> wrote:
>
>
> 👉📄 Download article proof <https://raw.githubusercontent.com/openjournals/joss-papers/joss.04611/joss.04611/10.21105.joss.04611.pdf> 📄 View article proof on GitHub <https://github.com/openjournals/joss-papers/blob/joss.04611/joss.04611/10.21105.joss.04611.pdf> 📄 👈
>
