### 2022-10-04 f2794be
Finish algorithm stability example 🙌🏻
