### 2022-10-04 4c93031
Add some docs explaining how to "analyze FS algorithm stability" (#83)
* [WIP] Analyzing algorithm stability recipe

* Finish stability example

* Reformat & imports

* Add note for relative vs absolute paths

* Algorithm stability recipe 📈 First version

* Finish algorithm stability example 🙌🏻
