+++ website/docs/_recipes/algorithm-stability.md
@@ -1,1 +1,0 @@
-# Analyze algorithm stability
