+++ examples/algorithm-stability-yaml/conf/ranker/boruta.yaml
@@ -1,0 +1,11 @@
+name: Boruta
+estimator:
+  _target_: boruta.boruta_py.BorutaPy
+  estimator:
+    _target_: sklearn.ensemble.RandomForestClassifier
+  n_estimators: auto
+_estimator_type: classifier
+multioutput: false
+estimates_feature_importances: false
+estimates_feature_support: true
+estimates_feature_ranking: true
