+++ examples/algorithm-stability-yaml/conf/metrics/stability_nogueira.yaml
@@ -1,0 +1,3 @@
+# @package metrics
+ranking_scores:
+  _target_: benchmark.StabilityNogueira
