+++ examples/algorithm-stability-yaml/conf/my_config.yaml
@@ -1,0 +1,11 @@
+defaults:
+  - base_pipeline_config
+  - _self_
+  - override dataset: synclf_hard
+  - override validator: knn
+  - override /callbacks:
+      - to_sql
+  - override /metrics:
+      - stability_nogueira
+
+n_bootstraps: 10
