### 2022-10-04 1402a47
Small-fix-in-algorithm-stability-plus-upgrade-docusaurus (#84)
