### 2022-10-05 8f448d0
Working example for running on AWS
