+++ examples/running-on-aws-using-ray/conf/ranker/mutual_info.yaml
@@ -1,0 +1,6 @@
+name: Mutual Info
+estimator:
+  _target_: benchmark.MutualInfoClassifier
+_estimator_type: classifier
+multioutput: false
+estimates_feature_importances: true
