### 2022-10-05 8307d33
Fix & refactor final AWS example
