### 2022-10-05 423cc59
Do actually stop the cluster
