### 2022-10-10 #4611 @estefaniatalavera
@dunnkers @diehlpk

This is a useful package for the community.  I would suggest the following improvement to the document:

- The first sentence of the statement of need mentions that FS and FR are extensively researched topics within machine learning, but the reference you include is from 2003. Aren’t there more recent works that present an overview of this field?
- If this is the first of such tools, clearly state it. If not, include a comparison with such other tools.
- If page limit allows it, this document would be more efficient into attracting new users by describing terminology and technology used. For instance, terms and concepts such as dataset adapters, bootstrap, AWS, among others could be described. Hydra was correctly presented.
- The document does not use numbering for the sections but they are referenced like that. You should fix that. For instance, you refer to section 3.2 in dataset adapters.
- The “Analyze algorithm stability” is mentioned but not described.
- I suggest you use the images equally in both the paper and documentation website.
- I miss a section on how the results look like and how they can be visualised or interpreted by the user of such package.

Improvements documentation website:
- Use verbs in present form. For example, in https://dunnkers.com/fseval/docs/the-pipeline/ pipeline step 4.
