### 2022-10-12 a549eb5
Recipe/running on aws (#86)
* Working example for running on AWS

* Fix & refactor final AWS example

* Do actually stop the cluster

* Running on AWS: docs 📚

* Running on AWS zip
