+++ examples/running-on-aws-using-ray/conf/my_config.yaml
@@ -1,0 +1,13 @@
+defaults:
+  - base_pipeline_config
+  - _self_
+  - override dataset: synthetic
+  - override validator: knn
+  - override /callbacks:
+      - to_sql
+  - override hydra/launcher: custom_ray_aws
+
+n_bootstraps: 1
+callbacks:
+  to_sql:
+    url: sqlite:////home/ubuntu/results/results.sqlite # any well-defined database URL
