### 2022-10-12 afdde2e
Install Node.js in the Devcontainer (#87)
Making it easier to develop the documentation website as well
