### 2022-10-12 5bee6bb
Updated schematic according to paper Now the schematic shown on the documentation page is on-par with whatever is shown in the JOSS paper submission
