### 2022-10-12 #4611 @dunnkers
Dear @estefaniatalavera @mcasl @diehlpk, I worked hard to process more feedback. These are some of the changes on the repo side of things (package and documentation):

- Installation instructions were improved (https://github.com/dunnkers/fseval/pull/81)
- 2 new recipes
  -  analyzing algorithm stability (https://github.com/dunnkers/fseval/pull/83)
  - running `fseval` on AWS (https://github.com/dunnkers/fseval/pull/86)
- Pipeline schematic was updated on the website (https://github.com/dunnkers/fseval/pull/88)

For the paper, I have an open PR here:

- Improve paper (https://github.com/dunnkers/fseval/pull/85)

Which is to be shortly reviewed by my peer-submitters.

That said, we have been (and are actively) working on processing the feedback of both @mcasl and @estefaniatalavera.
