### 2022-10-14 2f0a04b
Cherry-pick Ahmad's changes 🍒 From https://github.com/amjams/fseval/tree/patch-1, https://github.com/amjams/fseval/tree/patch-2 and https://github.com/amjams/fseval/tree/patch-3 🙌🏻.
