+++ paper.bib
@@ -315,3 +315,13 @@ Publisher: ACM},
   year={2020},
   publisher={Elsevier}
 }
+
+@article{venkatesh2019review,
+  title={A review of feature selection and its methods},
+  author={Venkatesh, B and Anuradha, J},
+  journal={Cybernetics and Information Technologies},
+  volume={19},
+  number={1},
+  pages={3--26},
+  year={2019}
+}
