### 2022-10-14 d5d394a
Process feedback from both reviewers into paper (#85)
* Process feedback from both reviewers into paper

* Cherry-pick Ahmad's changes 🍒
From https://github.com/amjams/fseval/tree/patch-1, https://github.com/amjams/fseval/tree/patch-2 and https://github.com/amjams/fseval/tree/patch-3 🙌🏻.
