+++ paper.bib
@@ -218,3 +218,110 @@ Publisher: ACM},
 	pages = {1--54},
 	file = {Full Text PDF:/Users/dunnkers/Zotero/storage/CXXAK82C/Nogueira et al. - 2018 - On the Stability of Feature Selection Algorithms.pdf:application/pdf},
 }
+
+@article{reis_featsel_2017,
+	title = {featsel: A framework for benchmarking of feature selection algorithms and cost functions},
+	journal = {SoftwareX},
+	volume = {6},
+	pages = {193-197},
+	year = {2017},
+	issn = {2352-7110},
+	doi = {https://doi.org/10.1016/j.softx.2017.07.005},
+	url = {https://www.sciencedirect.com/science/article/pii/S2352711017300286},
+	author = {Marcelo S. Reis and Gustavo Estrela and Carlos Eduardo Ferreira and Junior Barrera},
+	keywords = {Feature selection, Benchmarking, Boolean lattice, Combinatorial optimization},
+	abstract = {In this paper, we introduce featsel, a framework for benchmarking of feature selection algorithms and cost functions. This framework allows the user to deal with the search space as a Boolean lattice and has its core coded in C++ for computational efficiency purposes. Moreover, featsel includes Perl scripts to add new algorithms and/or cost functions, generate random instances, plot graphs and organize results into tables. Besides, this framework already comes with dozens of algorithms and cost functions for benchmarking experiments. We also provide illustrative examples, in which featsel outperforms the popular Weka workbench in feature selection procedures on data sets from the UCI Machine Learning Repository.}
+}
+
+@article{cilia2019experimental,
+  title={An experimental comparison of feature-selection and classification methods for microarray datasets},
+  author={Cilia, Nicole Dalia and De Stefano, Claudio and Fontanella, Francesco and Raimondo, Stefano and Scotto di Freca, Alessandra},
+  journal={Information},
+  volume={10},
+  number={3},
+  pages={109},
+  year={2019},
+  publisher={MDPI}
+}
+ 
+@article{sun2019comparison,
+  title={Comparison of feature selection methods and machine learning classifiers for radiomics analysis in glioma grading},
+  author={Sun, Pan and Wang, Defeng and Mok, Vincent Ct and Shi, Lin},
+  journal={IEEE Access},
+  volume={7},
+  pages={102010--102020},
+  year={2019},
+  publisher={IEEE}
+}
+
+@article{cilia2019experimental,
+  title={An experimental comparison of feature-selection and classification methods for microarray datasets},
+  author={Cilia, Nicole Dalia and De Stefano, Claudio and Fontanella, Francesco and Raimondo, Stefano and Scotto di Freca, Alessandra},
+  journal={Information},
+  volume={10},
+  number={3},
+  pages={109},
+  year={2019},
+  publisher={MDPI}
+}
+ 
+@article{sun2019comparison,
+  title={Comparison of feature selection methods and machine learning classifiers for radiomics analysis in glioma grading},
+  author={Sun, Pan and Wang, Defeng and Mok, Vincent Ct and Shi, Lin},
+  journal={IEEE Access},
+  volume={7},
+  pages={102010--102020},
+  year={2019},
+  publisher={IEEE}
+}
+ 
+ @article{tohka2016comparison,
+  title={Comparison of feature selection techniques in machine learning for anatomical brain MRI in dementia},
+  author={Tohka, Jussi and Moradi, Elaheh and Huttunen, Heikki},
+  journal={Neuroinformatics},
+  volume={14},
+  number={3},
+  pages={279--296},
+  year={2016},
+  publisher={Springer}
+}
+ 
+ 
[…]
