### 2022-10-14 9461a0a
Add Developer instructions By two means: 1. A CONTRIBUTING.md file 2. A "Developer Guide" page on the documentation website
