### 2022-10-14 c797d1b
Add Developer instructions (#89)
* Add Developer instructions
By two means:
1. A CONTRIBUTING.md file
2. A "Developer Guide" page on the documentation website

* Fix contributing link
