+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+# Contributing to `fseval`
+Awesome that you are looking to contribute to `fseval`! Be it just a small spelling-correction, a bugfix, or a bigger and more structural change; all contributions are welcome. Let's make the process as easy and transparent for you as possible.
+
+Please see the [developer guide](https://dunnkers.com/fseval/docs/contributing/) on the website.
+
+## Pull Requests
+We welcome your pull requests.
+
+1. Fork the repo and create your feature branch from `master`.
+2. If you've added code add suitable tests.
+3. If you've changed APIs, update the documentation.
+4. Ensure the test suite and lint pass.
+
+## Issues
+We use GitHub issues to track public bugs. Please ensure your description is
+clear and has sufficient instructions to be able to reproduce the issue.
+
+## License
+By contributing to hydra, you agree that your contributions will be licensed
+under the LICENSE file in the root directory of this source tree.
