+++ assets/js/935f2afb.45cc3e6e.js
@@ -1,0 +1,1 @@
+"use strict";(self.webpackChunkwebsite=self.webpackChunkwebsite||[]).push([[53],{1109:function(e){e.exports=JSON.parse('{"pluginId":"default","version":"current","label":"Next","banner":null,"badge":false,"noIndex":false,"className":"docs-version-current","isLast":true,"docsSidebars":{"tutorialSidebar":[{"type":"link","label":"Motivation","href":"/fseval/docs/motivation","docId":"motivation"},{"type":"link","label":"Getting started","href":"/fseval/docs/quick-start","docId":"quick-start"},{"type":"link","label":"The pipeline","href":"/fseval/docs/the-pipeline","docId":"the-pipeline"},{"type":"link","label":"fseval.main","href":"/fseval/docs/main","docId":"main"},{"type":"category","label":"fseval.config","collapsible":true,"collapsed":true,"items":[{"type":"link","label":"fseval.config.PipelineConfig","href":"/fseval/docs/config/PipelineConfig","docId":"config/PipelineConfig"},{"type":"link","label":"fseval.config.DatasetConfig","href":"/fseval/docs/config/DatasetConfig","docId":"config/DatasetConfig"},{"type":"link","label":"fseval.config.CrossValidatorConfig","href":"/fseval/docs/config/CrossValidatorConfig","docId":"config/CrossValidatorConfig"},{"type":"link","label":"fseval.config.ResampleConfig","href":"/fseval/docs/config/ResampleConfig","docId":"config/ResampleConfig"},{"type":"link","label":"fseval.config.EstimatorConfig","href":"/fseval/docs/config/EstimatorConfig","docId":"config/EstimatorConfig"},{"type":"link","label":"fseval.config.StorageConfig","href":"/fseval/docs/config/StorageConfig","docId":"config/StorageConfig"},{"type":"link","label":"Metrics","href":"/fseval/docs/config/metrics","docId":"config/metrics"},{"type":"link","label":"Callbacks","href":"/fseval/docs/config/callbacks","docId":"config/callbacks"}]},{"type":"category","label":"fseval.types","collapsible":true,"collapsed":true,"items":[{"type":"link","label":"fseval.types.CacheUsage","href":"/fseval/docs/types/CacheUsage","docId":"types/CacheUsage"},{"type":"link","label":"fseval.types.Task","href":"/fseval/docs/types/Task","docId":"types/Task"}]},{"type":"category","label":"Recipes","collapsible":true,"collapsed":true,"items":[{"type":"link","label":"Analyze algorithm stability","href":"/fseval/docs/recipes/algorithm-stability","docId":"recipes/algorithm-stability"},{"type":"link","label":"Comparing Feature Selectors","href":"/fseval/docs/recipes/comparing-feature-selectors","docId":"recipes/comparing-feature-selectors"},{"type":"link","label":"Running on AWS","href":"/fseval/docs/recipes/running-on-aws","docId":"recipes/running-on-aws"}]}]},"docs":{"config/callbacks":{"id":"config/callbacks","title":"Callbacks","description":"Callbacks are responsible for storing the experiment config and results. Any callback can be used on its own, but multiple callbacks can also be used at the same time.","sidebar":"tutorialSidebar"},"config/CrossValidatorConfig":{"id":"config/CrossValidatorConfig","title":"fseval.config.CrossValidatorConfig","description":"Cross Validation is used to improve the reliability of the ranking and validation results. A CV method can be defined like so:","sidebar":"tutorialSidebar"},"config/DatasetConfig":{"id":"config/DatasetConfig","title":"fseval.config.DatasetConfig","description":"Configures a dataset, to be used in the pipeline. Can be loaded from variou
[…]
