+++ website/docs/recipes/comparing-feature-selectors.md
@@ -1,0 +1,617 @@
+# Comparing Feature Selectors
+Hi! You want to compare the performance of multiple feature selectors? This is an example Notebook, showing you how to do such an analysis. 
+
+## Prerequisites
+
+We are going to use more or less the same configuration as we did in the [Quick start](../../quick-start) example, but then with more Feature Selectors. Again, start by downloading the example project: [comparing-feature-selectors.zip](pathname:///fseval/zipped-examples/comparing-feature-selectors.zip)
+
+### Installing the required packages
+
+Now, let's install the required packages. Make sure you are in the `comparing-feature-selectors` folder, containing the `requirements.txt` file, and then run the following:
+
+```
+pip install -r requirements.txt
+```
+
+## Running the experiment
+
+Run the following command to start the experiment:
+
+```
+python benchmark.py --multirun ranker="glob(*)" +callbacks.to_sql.url="sqlite:////tmp/results.sqlite
+```
+
+## Analyzing the results
+
+There now should exist a `.sqlite` file at this path: `/tmp/results.sqlite`:
+
+    ```
+    $ ls -al /tmp/results.sqlite
+    -rw-r--r-- 1 vscode vscode 20480 Sep 21 08:16 /tmp/results.sqlite
+    ```
+
+Is that the case? Then let's now analyze the results! 📈
+
+We will install `plotly-express`, so we can make nice plots later.
+
+
+```python
+%pip install plotly-express nbconvert --quiet
+```
+
+
+Next, let's find a place to store our results to. In this case, we choose to store it in a local SQLite database, located at `/tmp/results.sqlite`.
+
+
+```python
+import os
+
+con: str = "sqlite:////tmp/results.sqlite"
+con
+```
+
+
+
+
+    'sqlite:////tmp/results.sqlite'
+
+
+
+Now, we can read the `experiments` table.
+
+
+```python
+import pandas as pd
+
+experiments: pd.DataFrame = pd.read_sql_table("experiments", con=con, index_col="id")
+experiments
+```
+
+
+
+
+<div>
+<table border="1" class="dataframe">
+  <thead>
+    <tr>
+      <th></th>
+      <th>dataset</th>
+      <th>dataset/n</th>
+      <th>dataset/p</th>
+      <th>dataset/task</th>
+      <th>dataset/group</th>
+      <th>dataset/domain</th>
+      <th>ranker</th>
+      <th>validator</th>
+      <th>local_dir</th>
+      <th>date_created</th>
+    </tr>
+    <tr>
+      <th>id</th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+      <th></th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <th>3lllxl48</th>
+      <td>My synthetic dataset</td>
+      <td>10000</td>
+      <td>20</td>
+      <td>classification</td>
+      <td>None</td>
+      <td>None</td>
+      <td>ANOVA F-value</td>
+      <td>k-NN</td>
+      <td>/workspaces/fseval/examples/comparing-feature-...</td>
+      <td>2022-10-22 14:28:27.506838</td>
+    </tr>
+    <tr>
+      <th>1944ropg</th>
+      <td>My synthetic dataset</td>
[…]
