+++ examples/comparing-feature-selectors/conf/ranker/stability_selection.yaml
@@ -1,0 +1,10 @@
+name: Stability Selection
+estimator:
+  _target_: benchmark.StabilitySelection
+  base_estimator:
+    _target_: sklearn.linear_model.LogisticRegression
+    penalty: l2
+  bootstrap_func: stratified
+_estimator_type: classifier
+estimates_feature_importances: true
+estimates_feature_support: true
