### 2022-10-22 #4611 @dunnkers
Update: added a recipe called "[Comparing Feature Selectors](https://dunnkers.com/fseval/docs/recipes/comparing-feature-selectors/)", explaining how to run a benchmark of bigger size and analyse the results in a Jupyter-notebook style explanation. Shows the experiment results using both a table and a plot.

- ( https://github.com/dunnkers/fseval/pull/90 )
