### 2022-10-31 #4611 @dunnkers
> @dunnkers you mentioned that you were waiting by your peer-submitters, to review the updated version of the paper. Could you confirm if this has already happened?

Hi @estefaniatalavera. Yes this has happened (mentioned in [this comment](https://github.com/openjournals/joss-reviews/issues/4611#issuecomment-1276349953) ✅). Cheers! 🌞
