### 2022-11-08 517dee6
Add CodeQL workflow for GitHub code scanning
