### 2022-11-17 #4611 @mcasl
I still fail to see any advance in some of the topics that I previously mentioned:
-  A statement of need: There's no statement of need in the documentation, only in the paper.
- Installation instructions: There is no list of dependencies
- Functionality documentation: No guidelines, ..., have been added yet.
