+++ website/docs/motivation.md
@@ -4,7 +4,7 @@ sidebar_position: 0
 
 # Motivation
 
-fseval helps you benchmark **Feature Selection** and **Feature Ranking** algorithms. Any algorithm that ranks features in importance.
+`fseval` helps you benchmark **Feature Selection** and **Feature Ranking** algorithms. Any algorithm that ranks features in importance.
 
 It comes useful if you are one of the following types of users:
 1. **Feature Selection / Feature Ranker algorithm authors**. You are the author of a novel Feature Selection algorithm. Now, you have to prove the performance of your algorithm against other competitors. Therefore, you are going to run a large-scale benchmark. Many authors, however, spend much time rewriting similar pipelines to benchmark their algorithms. fseval helps you run benchmarks in a structured manner, on supercomputer clusters or on the cloud.
@@ -12,28 +12,17 @@ It comes useful if you are one of the following types of users:
 1. **Machine Learning practitioners**. You have a dataset and want to find out with exactly what features your models will perform best. You can use fseval to try multiple Feature Selection or Feature Ranking algorithms.
 
 ## Statement of Need
-So why would you need the help of `fseval`? Let's take a look.
 
-Feature Selection (FS) and Feature Ranking (FR) are 
-extensively researched topics within machine learning
-[@venkatesh2019review; @guyon_introduction_2003]. FS methods determine subsets of relevant
-features in a dataset, whereas FR methods rank the features in a dataset
-relative to each other in terms of their relevance. When a new FS or FR
-method is developed, a benchmarking scheme is necessary to empirically
-validate its effectiveness. Often, the benchmark is conducted as
-follows: features are ranked by importance, then the predictive quality
-of the feature subsets containing the top ranked features is evaluated
-using a validation estimator. Some studies let the competing FS or FR
-algorithms pick out a fixed number of top $k$ features and validate the
-performance of that feature subset
-[@roffo_infinite_2015; @zhao_searching_2007; @bradley_feature_1998],
-whilst others evaluate multiple subsets of increasing cardinality
-containing the highest ranked features
-[@wojtas_feature_2020; @bennasar_feature_2015; @gu_generalized_2012; @peng_feature_2005; @kira_feature_1992; @almuallim_learning_1991]. 
-FS algorithms that only make a binary prediction on which features to
-keep, are always evaluated in the former way.
+<details>
+<summary>
+About benchmarking Feature Selection and Feature Ranking algorithms
+</summary>
 
-There is a clear case for performing Feature Selection, as it has been shown to improve classification performance in many tasks, especially those with a large number of features and limited observations. In those applications, it is difficult to determine which FS method is suitable in the general case. Therefore, large empirical comparisons of several FS methods and classifiers are routinely performed. For instance, in microarray data [@cilia2019experimental], medical imaging [@sun2019comparison; @tohka2016comparison; @ashok2016comparison], and text classification [@liu2017multi; @kou2020evaluation]. Therefore, it is valuable to find out emperically which FR- or FS method works best. This requires running a benchmark to do so.
[…]
