+++ .github/workflows/python-app.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install mypy pytest pytest-cov black
+          pip install mypy==0.982 pytest pytest-cov black
           pip install -r requirements.txt
       - name: Lint with mypy
         run: |
