### 2022-11-18 bd45bed
Fix mypy error
