### 2022-11-18 f49ba8b
Add statement of need to docs (#92)
* Put in Statement of Need from paper

* Finish adding statement of need 👍🏻

* Fix mypy version

* Fix mypy *

* Fix mypy error

* Ignore mypy error
