+++ .devcontainer/requirements.txt
@@ -1,4 +1,4 @@
-mypy
+mypy==0.982
 pytest
 pytest-cov
 black
