### 2022-11-18 #4611 @dunnkers
> I still fail to see any progress in some of the topics that I previously mentioned:
>
> * A statement of need: There's no statement of need in the documentation, only in the paper.
> * Installation instructions: There is no list of dependencies
> * Functionality documentation: No guidelines have been added yet.
>
> The other issues have been appropriately addressed and the checklist reflects the changes.

Hi @mcasl! Thanks for the feedback.

See below a response to each of your points 🙂:

## Statement of need
> * A statement of need: There's no statement of need in the documentation, only in the paper.

✓ Added in:
- https://github.com/dunnkers/fseval/pull/92 (see [website](https://dunnkers.com/fseval/docs/motivation/#statement-of-need))

## Dependency list
> * Installation instructions: There is no list of dependencies

The checklist says the following:

> Installation instructions: Is there a clearly-stated list of dependencies? **Ideally these should be handled with an automated package management solution.**

Saying; the dependencies should _ideally_ be handled by an automated package management solution. This is the case for `fseval`. We are using a [`setup.py` ](https://github.com/dunnkers/fseval/blob/master/setup.py#L30-L40) file which lists all the required dependencies, which are then automatically installed when the user installs the package.

Furthermore, there exists a [`requirements.txt`](https://github.com/dunnkers/fseval/blob/master/requirements.txt) file, listing all dependencies in such a way that they can automatically consumed and installed by Python's package manager, `pip`. This is referred to in the "Installation from source" [installation guide](https://github.com/dunnkers/fseval#install).

Additionally, all dependencies required for developing are listed in another [`requirements.txt`](https://github.com/dunnkers/fseval/blob/master/.devcontainer/requirements.txt) file. This is referred to in the [Developer Guide](https://dunnkers.com/fseval/docs/contributing/#option-b-installing-the-dependencies-manually). If the user prefers to use our containerized solution, we also packaged up a Docker image for them which can be used by spinning up the project's [Devcontainer](https://dunnkers.com/fseval/docs/contributing/#option-a-using-a-devcontainer).

## Community guidelines
> * Functionality documentation: No guidelines have been added yet.

Looking at your checklist I assume the guidelines mentioned refer to the _community_ guidelines:

> * ❌ **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The community guidelines were added in this Pull Request:

- [x] Added Developer guidelines for contributing (https://github.com/dunnkers/fseval/pull/89)

That is, guidelines for the community were added both to the documentation website and [in the repo](https://github.com/dunnkers/fseval/blob/master/CONTRIBUTING.md). I could see, however, that the website was not yet re-deployed after the changes 😬. I made a re-deployment just now so the [Developer Guide](https://dunnkers.com/fseval/docs/contributing/) is now online.
