+++ paper.bib
@@ -234,48 +234,31 @@ Publisher: ACM},
 }
 
 @article{cilia2019experimental,
-  title={An experimental comparison of feature-selection and classification methods for microarray datasets},
-  author={Cilia, Nicole Dalia and De Stefano, Claudio and Fontanella, Francesco and Raimondo, Stefano and Scotto di Freca, Alessandra},
-  journal={Information},
-  volume={10},
-  number={3},
-  pages={109},
-  year={2019},
-  publisher={MDPI}
+  author = {Cilia, Nicole Dalia and De Stefano, Claudio and Fontanella, Francesco and Raimondo, Stefano and Scotto di Freca, Alessandra},
+  title = {An Experimental Comparison of Feature-Selection and Classification Methods for Microarray Datasets},
+  journal = {Information},
+  volume = {10},
+  year = {2019},
+  number = {3},
+  article-number = {109},
+  url = {https://www.mdpi.com/2078-2489/10/3/109},
+  issn = {2078-2489},
+  abstract = {In the last decade, there has been a growing scientific interest in the analysis of DNA microarray datasets, which have been widely used in basic and translational cancer research. The application fields include both the identification of oncological subjects, separating them from the healthy ones, and the classification of different types of cancer. Since DNA microarray experiments typically generate a very large number of features for a limited number of patients, the classification task is very complex and typically requires the application of a feature-selection process to reduce the complexity of the feature space and to identify a subset of distinctive features. In this framework, there are no standard state-of-the-art results generally accepted by the scientific community and, therefore, it is difficult to decide which approach to use for obtaining satisfactory results in the general case. Based on these considerations, the aim of the present work is to provide a large experimental comparison for evaluating the effect of the feature-selection process applied to different classification schemes. For comparison purposes, we considered both ranking-based feature-selection techniques and state-of-the-art feature-selection methods. The experiments provide a broad overview of the results obtainable on standard microarray datasets with different characteristics in terms of both the number of features and the number of patients.},
+  doi = {10.3390/info10030109}
 }
  
 @article{sun2019comparison,
-  title={Comparison of feature selection methods and machine learning classifiers for radiomics analysis in glioma grading},
   author={Sun, Pan and Wang, Defeng and Mok, Vincent Ct and Shi, Lin},
   journal={IEEE Access},
-  volume={7},
-  pages={102010--102020},
+  title={Comparison of Feature Selection Methods and Machine Learning Classifiers for Radiomics Analysis in Glioma Grading},
   year={2019},
-  publisher={IEEE}
+  volume={7},
+  number={},
+  pages={102010-102020},
+  doi={10.1109/ACCESS.2019.2928975}
 }
 
-@article{cilia2019experimental,
-  title={An experimental comparison of feature-selection and classification methods for microarray datasets},
-  author={Cilia, Nicole Dalia and De Stefano, Claudio and Fontanella, Francesco and Raimondo, Stefano and Scotto di Freca, Alessandra},
-  journal={Information},
-  volume={10},
-  number={3},
-  pages={109},
-  year={2019},
-  publisher={MDPI}
-}
- 
-@article{sun2019comparison,
-  title={Comparison of feature selection methods and machine learning classifiers for radiomics analysis in glioma grading},
-  author={Sun, Pan and Wang, Defeng and Mok, Vincent Ct and Shi, Lin},
-  journal={IEEE Access},
-  volume={7},
-  pages={102010--102020},
-  year={2019},
-  publisher={IEEE}
-}
- 
- @article{tohka2016comparison,
+@article{tohka2016comparison,
   title={Comparison of feature selection techniques in machine learning for anatomical brain MRI in dementia},
[…]
