+++ paper.bib
@@ -226,7 +226,7 @@ Publisher: ACM},
 	pages = {193-197},
 	year = {2017},
 	issn = {2352-7110},
-	doi = {https://doi.org/10.1016/j.softx.2017.07.005},
+	doi = {10.1016/j.softx.2017.07.005},
 	url = {https://www.sciencedirect.com/science/article/pii/S2352711017300286},
 	author = {Marcelo S. Reis and Gustavo Estrela and Carlos Eduardo Ferreira and Junior Barrera},
 	keywords = {Feature selection, Benchmarking, Boolean lattice, Combinatorial optimization},
@@ -278,7 +278,7 @@ Publisher: ACM},
   pages = {323-339},
   year = {2017},
   issn = {0957-4174},
-  doi = {https://doi.org/10.1016/j.eswa.2017.03.042},
+  doi = {10.1016/j.eswa.2017.03.042},
   url = {https://www.sciencedirect.com/science/article/pii/S0957417417301951},
   author = {Yang Liu and Jian-Wu Bi and Zhi-Ping Fan},
   keywords = {Multi-class sentiment classification, Experimental comparison, Feature selection algorithms, Machine learning algorithms},
@@ -292,7 +292,7 @@ Publisher: ACM},
   pages = {105836},
   year = {2020},
   issn = {1568-4946},
-  doi = {https://doi.org/10.1016/j.asoc.2019.105836},
+  doi = {10.1016/j.asoc.2019.105836},
   url = {https://www.sciencedirect.com/science/article/pii/S1568494619306179},
   author = {Gang Kou and Pei Yang and Yi Peng and Feng Xiao and Yang Chen and Fawaz E. Alsaadi},
   keywords = {Feature selection, Text classification, MCDM, Small sample dataset},
@@ -301,7 +301,7 @@ Publisher: ACM},
 
 @article{venkatesh2019review,
   author = {B. Venkatesh and J. Anuradha},
-  doi = {doi:10.2478/cait-2019-0001},
+  doi = {10.2478/cait-2019-0001},
   url = {https://doi.org/10.2478/cait-2019-0001},
   title = {A Review of Feature Selection and Its Methods},
   journal = {Cybernetics and Information Technologies},
