### 2022-11-21 #4611 @dunnkers
> Generate a new release with all the changes you made during the review. Please post the version number here.
Released version **3.1.0** ✓ https://github.com/dunnkers/fseval/releases/tag/3.1.0

> Upload this release do Zenodo or FigShare and provide me with the DPI. Note that the authors and title of the Zenodo archive have to match with the paper.

DOI: 10.5281/zenodo.7343417

[Version 3.1.0 ](https://zenodo.org/record/7343417)10.5281/zenodo.734341, Nov 21, 2022

Is that okay? Not sure if the title matches; but not sure how to change
