+++ paper.bib
@@ -17,7 +17,7 @@
 }
 
 @inproceedings{kira_feature_1992,
-	title = {The {Feature} {Selection} {Problem}: {Traditional} {Methods} and a {New} {Algorithm}},
+	title = {The Feature Selection Problem: Traditional Methods and a New Algorithm},
 	shorttitle = {The {Feature} {Selection} {Problem}},
 	abstract = {For real-world concept learning problems, feature selection is important to speed up learning and to improve concept quality. We review and analyze past approaches to feature selection and note their strengths and weaknesses. We then introduce and theoretically examine a new algorithm Rellef which selects relevant features using a statistical method. Relief does not depend on heuristics, is accurate even if features interact, and is noise-tolerant. It requires only linear time in the number of given features and the number of training instances, regardless of the target concept complexity. The algorithm also has certain limitations such as nonoptimal feature set size. Ways to overcome the limitations are suggested. We also report the test results of comparison between Relief and other feature selection algorithms. The empirical results support the theoretical analysis, suggesting a practical approach to feature selection for real-world problems.},
 	booktitle = {{AAAI}},
@@ -48,7 +48,7 @@
 
 @inproceedings{roffo_infinite_2015,
 	address = {Santiago, Chile},
-	title = {Infinite {Feature} {Selection}},
+	title = {Infinite Feature Selection},
 	isbn = {978-1-4673-8391-2},
 	url = {http://ieeexplore.ieee.org/document/7410835/},
 	doi = {10.1109/ICCV.2015.478},
@@ -66,7 +66,7 @@
 }
 
 @inproceedings{almuallim_learning_1991,
-	title = {Learning {With} {Many} {Irrelevant} {Features}},
+	title = {Learning With Many Irrelevant Features},
 	abstract = {In many domains, an appropriate inductive bias is the MIN-FEATURES bias, which prefers consistent hypotheses definable over as few features as possible. This paper defines and studies this bias. First, it is shown that any learning algorithm implementing the MIN-FEATURES bias requires {\textbackslash}Theta(  1  ffl ln  1  ffi +  1  ffl [2  p  +  p ln n]) training examples to guarantee PAC-learning a concept having p relevant features out of n available features. This bound is only logarithmic in the number of irrelevant features. The paper also presents a quasi-polynomial time algorithm, FOCUS, which implements MIN-FEATURES. Experimental studies are presented that compare FOCUS to the ID3 and FRINGE algorithms. These experiments show that--- contrary to expectations---these algorithms do not implement good approximations of MIN-FEATURES. The coverage, sample complexity, and generalization performance of FOCUS is substantially better than either ID3 or FRINGE on learning problems where the MIN-FEATURE...},
 	booktitle = {In {Proceedings} of the {Ninth} {National} {Conference} on {Artificial} {Intelligence}},
 	publisher = {AAAI Press},
@@ -78,7 +78,7 @@
 }
 
 @article{gu_generalized_2012,
-	title = {Generalized {Fisher} {Score} for {Feature} {Selection}},
+	title = {Generalized {Fisher} Score for Feature Selection},
 	url = {http://arxiv.org/abs/1202.3725},
[…]
