### 2022-11-23 #4611 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
