### 2022-11-23 #4611 @dunnkers
> This DOI isn't resolving for me yet - I will leave this issue open for now, and come back and check again in a little while.  But other than that, the process is complete

👍. I also cannot get hold of the DOI link.

By the way: is it possible to have my name displayed as _Jeroen G. S. Overschie_ instead of _Jeroen G. s. Overschie_ (capitalised second names) on the JOSS website?

![image](https://user-images.githubusercontent.com/744430/203631793-39217cb0-8b5c-447e-b7cc-ce4eb704895e.jpeg)

That would be awesome. Thanks
