### 2022-11-23 3fd80c5
Add CITATION.cff file 📝📌
Paper was accepted and is now citable here: https://joss.theoj.org/papers/10.21105/joss.04611#. So let's add it to the repo 🚀.
