### 2022-11-24 #4611 @arfon
@xuanxu @tarleb – the name formatting looks good in the PDF here: https://github.com/openjournals/joss-papers/blob/f0d0a5e8b17e200bc09a0ecac00c59149ccba189/joss.04611/10.21105.joss.04611.pdf – is there a preferred way we can override the name in the front matter to accomplish this?
