### 2022-11-24 #4611 @arfon
DOI is now resolving but we'll need to re-reaccept when we figure out the capitalisation issue.
