### 2022-11-24 #4611 @tarleb
It seems to be limited to the HTML presentation; both the PDF and JATS have the correct capitalization. Not sure what might be causing this.
