### 2022-11-25 #4611 @xuanxu
Yes, the parsing was taking the second names as a dropping particle.
Reaccepting the paper after [this change](https://github.com/dunnkers/fseval/pull/95) should fix it.
